pallet_bitcoin::pallet

Type Alias Coins

pub type Coins<T> = StorageDoubleMap<_GeneratedPrefixForStorageCoins<T>, Identity, Txid, Identity, OutputIndex, Coin, OptionQuery>;

UTXO set.

(Txid, OutputIndex(vout), Coin)

Note: There is a special case in Bitcoin that the outputs in the genesis block are excluded from the UTXO set.

Storage type is [StorageDoubleMap] with key1 type Txid, key2 type OutputIndex and value type Coin.

Aliased Type

pub struct Coins<T>(/* private fields */);