Crate pallet_bitcoin

Bitcoin Pallet

This pallet is designed to be minimalist, containing only one storage item for maintaining the state of the UTXO (Unspent Transaction Output) set by processing the inputs and outputs of each Bitcoin transaction wrapped in Call::transact. There is no verification logic within the pallet, all validation work should be performed outside the runtime. This approach simplifies off-runtime execution, allowing for easier syncing performance optimization off chain.

Re-exports

pub use pallet::*;

Modules

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

Structs

Txid

Wrapper type for Bitcoin txid in runtime as bitcoin::Txid does not implement codec.

Functions

coin_storage_key

Returns the storage key for the referenced output.

coin_storage_prefix

Type Aliases

Vout

Transaction output index.